build-helper-maven-plugin

+++ java-program-design/src/main/java/chapter/three/four/two/OperatorsAndExpressions.java
@@ -8,7 +8,7 @@ public class OperatorsAndExpressions {
     }
 
     void shortCircuit2(int val) {
-        if ((val != 0) && (1 / val < Double.MAX_VALUE)) {
+        if ((val != 0) && ((double) 1 / val < Double.MAX_VALUE)) {
             System.out.println("val 的倒数为:" + 1 / val);
         } else {
             System.out.println("val 的倒数为无穷大");
